refactor: MemberRecordPeriodProvider 생성 및 적용
