+++ frontend/src/components/record/MemberRecords/MemberRecords.tsx
@@ -1,8 +1,5 @@
-import { Suspense } from 'react';
-
 import MemberRecordCalendar from '../MemberRecordCalendar/MemberRecordCalendar';
 import MemberRecordList from '../MemberRecordList/MemberRecordList';
-import MemberRecordListSkeleton from '../MemberRecordList/MemberRecordListSkeleton';
 
 type Props = {
   memberId: string;
@@ -12,11 +9,7 @@ type Props = {
 const MemberRecords = ({ memberId, viewMode }: Props) => {
   if (viewMode === 'calendar') return <MemberRecordCalendar />;
 
-  return (
-    <Suspense fallback={<MemberRecordListSkeleton />}>
-      <MemberRecordList memberId={memberId} />
-    </Suspense>
-  );
+  return <MemberRecordList memberId={memberId} />;
 };
 
 export default MemberRecords;
