+++ frontend/src/constants/record.ts
@@ -1,0 +1,6 @@
+export const PERIOD = {
+  week: '1주일',
+  oneMonth: '1개월',
+  threeMonth: '3개월',
+  entire: '전체 기간',
+} as const;
